// src/types/class.ts
export interface Class {

  id: string;

  organizationId: string;

  organizationName: string;

  curriculumId: string;

  curriculumName: string;

  classCode: string;

  className: string;

  displayOrder: number;

  isActive: boolean;

  createdBy?: string;

  updatedBy?: string;

  createdAt?: string;

  updatedAt?: string;
}